🚚 rename: Top10 -> Popular 로 파일명, 컴포넌트명 변경
