+++ grass-diary/src/components/index.ts
@@ -4,7 +4,7 @@ export { default as Button } from './Button/Button';
 export { default as Like } from './Like/Like';
 export { default as NormalLike } from './Like/NormalLike';
 export { default as Feed } from './Feed/Feed';
-export { default as Top10Feed } from './Feed/Top10Feed';
+export { default as PopularFeed } from './Feed/PopularFeed';
 export { default as Profile } from './Profile/Profile';
 export { default as MoodProfile } from './Profile/MoodProfile';
 export { default as ProtectedRoute } from './ProtectedRoute';
