+++ grass-diary/src/hooks/usePopularDiaries.ts
@@ -1,15 +1,15 @@
 import API from '@services/index';
 import { useQuery } from '@tanstack/react-query';
 
-const fetchTop10Diary = () => {
+const fetchPopularDiary = () => {
   return API.get('/shared/diaries/popular');
 };
 
-export const useTop10Diaries = () => {
+export const usePopularDiaries = () => {
   return useQuery({
     queryKey: ['top10'],
     queryFn: async (): Promise<IPopularResponse[]> => {
-      const res = await fetchTop10Diary();
+      const res = await fetchPopularDiary();
       return res.data;
     },
   });
